Remove all compilation warnings from engine's code
Verification:
1. Rebuild all engine projects (without deps) in all configurations.
   There should be only one warning:
   "Aligned.hpp(80): warning C4100: 'p' : unreferenced formal parameter",
   which is a VS compiler bug.
2. Run tests.

+++ nfEngine/nfCore/RenderCommand.cpp
@@ -163,7 +163,7 @@ void RenderCommandBuffer::Sort()
 
     //build array of pointers and manipulate them instead of RenderCommand structures
     int* pPointers = (int*)malloc(sizeof(int) * size);
-    for (int i = 0; i < size; i++)
+    for (size_t i = 0; i < size; i++)
         pPointers[i]  = i;
 
     int* pTmpPointers = (int*)malloc(sizeof(int) * size);
@@ -175,7 +175,7 @@ void RenderCommandBuffer::Sort()
     //reorganize
     RenderCommand* pTmpCommands = (RenderCommand*)_aligned_malloc(sizeof(RenderCommand) * size, 16);
     memcpy(pTmpCommands, &commands[0], sizeof(RenderCommand) * size );
-    for (int i = 0; i < commands.size(); i++)
+    for (size_t i = 0; i < commands.size(); i++)
     {
         commands[i] = pTmpCommands[pPointers[i]];
     }
